Seperate refresh from RuntimeService
